fixed a lot of things and prepped obj capture sims with sliding window averaging for collective radius

+++ microcontrollers/tether_bot/utils.h
@@ -16,7 +16,7 @@ float toDegrees(float radAngle);
 
 float mod(float x, float y);
 
-int sign(float x);
+float sign(float x);
 
 float vectorMagnitude(Matrix<2,1> vector);
